reorganize the structure of functional tests
